Refactor text appending functionality for global use
This commit extracts the functionality used to append user input and a
string of text, which was previously locked into the `GameWindow.svelte`
component, and places it into a separate utility file.

In order to create classes for the various entities in the game, we will
utilize `.ts` files, which cannot import functions from Svelte
componenents. In order to expose the text append functionality to these
classes, it is thus necessary to have a utility file from which both
`.ts` files and svelte components may import functions.