Add `Game` class and logic to application
This commit adds the `Game` class, allowing for the necessary logic to
create and interact with the world.

When the component mounts, a `Game` instance will be created, containing
the necessary world data (rooms, events, items, etc.). Users can now
enter commands in the input menu to interact with the world, which
responds dynamically.