Add `campfire` room and ability to alter examination flavor text
Adds logic for the campfire room, as well as associated "Matches" item
and art.

Additionally, a new method has been added to allow event to change the
flavor text of the examination features (necessary in the case of the
campfire). This allows the game to respond to users interacting with the
environment (such as taking an item).