Add item recognition in user input
This commit adds functionality to recognize if the user types in an item
in their inventory. If so, the game then asks the user what they wish to
do with this item.

This will be helpful once item interaction is added, since it will
indicate to the user that there is something that can be done with that
item, but perhaps they haven't provided the correct command or are in
the wrong location to do so.

Additional small changes to the world flavor text as well.