Add barebones item tracker to main screen
This commit adds a piece of text to the side of the game screen that
tells the user how many of the total available items they have found. I
figured it would be helpful to prompt the user to explore more if they
knew that there were still items laying around (since I think its easy
to get discouraged with a text RPG).